Collapse directive
this is a directive to make adding a collapse affect tied to scroll views easier to add. there are some layout updates required for it. However they are minor. simply add the collapse direcrive to the item you want to "collapse" then pass in the list item to watch the scrolling events on.