Parallax v2
The previous implementation didn't seem as smooth as it should. So instead of chaging the heigh of the top container, this adjusts the y postion based on the veritcal offset of the scroll. the entire scroll view scrolls the same, but as the item scrolls down the header is pushed down below the other items in the gridview creating the parallax effect. I used animations in ios to help create a smother effect and it looks good. however in Andoid the animation caused a very delayed effect. so instead it just adjuts the Y position without any animation.